Cleanup sample
