+++ framework/src/modules/chain/transactions/broadcaster.js
@@ -276,4 +276,6 @@ class Broadcaster {
 	}
 }
 
-module.exports = Broadcaster;
+module.exports = {
+	Broadcaster,
+};
