+++ framework/src/modules/chain/transactions/index.js
@@ -15,7 +15,14 @@
 'use strict';
 
 const { Transactions } = require('./transactions');
+const { Transaction } = require('./transaction');
+const { TransactionPool } = require('./transaction_pool');
+const { sort, sortBy } = require('./sort');
 
 module.exports = {
 	Transactions,
+	TransactionPool,
+	Transaction,
+	sortBy,
+	sort,
 };
