+++ framework/src/modules/chain/transactions/transaction.js
@@ -203,4 +203,6 @@ class Transaction {
 	}
 }
 
-module.exports = Transaction;
+module.exports = {
+	Transaction,
+};
